requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apify_client==1.12.0
apify_shared==1.4.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blis==1.3.0
boto3==1.38.27
botocore==1.38.27
brevo-python==1.2.0
bs4==0.0.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cobble==0.1.4
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
fastapi==0.116.1
fonttools==4.59.0
frozenlist==1.7.0
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
helpers==0.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-core==0.3.72
langchain-openai==0.3.28
langcodes==3.5.0
langgraph==0.6.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.0
langgraph-sdk==0.2.0
langsmith==0.4.8
language_data==1.3.0
logtail-python==0.3.3
lxml==6.0.0
Mako==1.3.10
mammoth==1.9.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
newrelic==10.15.0
nltk==3.9.1
numpy==2.2.6
openai==1.97.1
opencv-python==4.12.0.88
orjson==3.11.1
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
postgrest==1.1.1
preshed==3.0.10
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
PyMySQL==1.1.1
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
s3transfer==0.13.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
